+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod detector;
-pub mod homography;
-pub mod quad;
+pub mod image_util;
+// pub mod quad;
+pub mod math_util;
 pub mod tag_families;
 
 pub use tag_families::TagFamily;
